requirements.txt
streamlit>=1.45.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
